Refactor code in mapList directory
