+++ app/src/androidTest/java/com/sdpteam/connectout/event/EventCreatorViewModelTest.java
@@ -7,7 +7,7 @@ import static org.junit.Assert.assertTrue;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
-import com.sdpteam.connectout.map.GPSCoordinates;
+import com.sdpteam.connectout.mapList.map.GPSCoordinates;
 import com.sdpteam.connectout.utils.LiveDataTestUtil;
 
 import org.junit.Test;
