+++ app/src/main/java/com/sdpteam/connectout/event/EventActivity.java
@@ -12,17 +12,16 @@ import androidx.lifecycle.ViewModelProvider;
 
 import com.sdpteam.connectout.R;
 import com.sdpteam.connectout.WithFragmentActivity;
-import com.sdpteam.connectout.map.GPSCoordinates;
-import com.sdpteam.connectout.map.MapModelManager;
-import com.sdpteam.connectout.map.MapViewFragment;
-import com.sdpteam.connectout.map.MapViewModel;
-import com.sdpteam.connectout.map.MapViewModelFactory;
+import com.sdpteam.connectout.mapList.map.GPSCoordinates;
+import com.sdpteam.connectout.mapList.MapListModelManager;
+import com.sdpteam.connectout.mapList.MapListViewModel;
+import com.sdpteam.connectout.mapList.map.MapViewFragment;
+import com.sdpteam.connectout.mapList.MapListViewModelFactory;
 import com.sdpteam.connectout.profile.ProfileID;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
-import java.util.Set;
 
 public class EventActivity extends WithFragmentActivity {
 
@@ -50,18 +49,20 @@ public class EventActivity extends WithFragmentActivity {
     }
 
     private void initMapFragment(Event event) {
-        final Fragment map = new MapViewFragment();
-        replaceFragment(map, R.id.event_fragment_container);
 
         // Model that returns a singleton of the event as the event list
-        final MapModelManager mapModel = () -> {
+        final MapListModelManager mapModel = (String filteredAttribute, String expectedValue) -> {
             final MutableLiveData<List<Event>> data = new MutableLiveData<>();
             data.setValue(Collections.singletonList(event)); // only show one marker in the map
             return data;
         };
+        // Implicitly instantiating MapListViewModel to use that instance back in MapViewFragment
+        final MapListViewModel mapViewModel = new ViewModelProvider(this, new MapListViewModelFactory(mapModel)).get(MapListViewModel.class);
 
-        // Implicitly instantiating MapViewModel to use that instance back in MapViewFragment
-        new ViewModelProvider(this, new MapViewModelFactory(mapModel)).get(MapViewModel.class);
+
+        final MapViewFragment map = new MapViewFragment(mapViewModel);
+        replaceFragment(map, R.id.event_fragment_container);
+        map.spotsOn(event);
     }
 
     private Event getEvent() {
